[CustomOp] Initial draft of lookup in new class hierarchy
